Changed rssi and snr fields in readings to type double

+++ src/com/trandonsystems/tekelek/database/UnitDAL.java
@@ -86,9 +86,9 @@ public class UnitDAL {
 			spStmt.setInt(17, reading.serviceDoorOpen ? 1 : 0);
 			spStmt.setInt(18, reading.flapStuckOpen ? 1 : 0);
 			spStmt.setInt(19, reading.nbIoTSignalStrength);
-			spStmt.setInt(20, reading.rssi);
-			spStmt.setInt(21, reading.src);
-			spStmt.setInt(22, reading.snr);
+			spStmt.setDouble(20, reading.rssi);
+			spStmt.setDouble(21, reading.src);
+			spStmt.setDouble(22, reading.snr);
 			spStmt.setInt(23, reading.ber);
 
 			// Convert java.time.Instant to java.sql.timestamp
